Don't include the feed tiddler itself to avoid infinite recursion
It's still possible to create an infinite recursion
by making two feeds include each other. This fix is still good to
have as it unblocks me to test the feed in a real world setting.

A real fix will entail detecting infinite recursion or
refusing to render tiddlers that include the atomfeed macro or
something else I'm not thinking of right now.