Prefix with tiddler title only when passing content through
When using a tiddler to render an atom feed, for example,
the title shouldn't be part of the response body.

While an atom feed tiddler is only a part of the potential use cases
of letting TiddlyWiki render the content before responding, it is
the use case that motivated the rendering mechanism. Until other
use cases emerge, optimizing for atom feeds seems fine.

Moreover, not including the title by default is more flexible,
as there's always the workaround of manually including the title
in the tiddler text. (On the other hand, if the title is included
by default, there's no way around it.)